separated writing functionality from the main file into separate classes and created an interface called Writer
