+++ src/writer/writer.cc
@@ -1,0 +1,7 @@
+// Copyright 2015 Akshay Dixit
+
+#include "writer.h"
+
+Writer::Writer(const std::string& outputFileName) {
+    outputFileBuffer.open(outputFileName);
+}
